\name{colsum_Mat}
\alias{colsum_Mat}
\title{
C++ function for column-wise numeric calculation in a matrix for a static binary undirected network.
}
\description{
The C++ function colsum_Mat() calculates the sum of each column in a matrix in a static binary undirected network.
}
\usage{
colsum_Mat(M)
}
\arguments{
  \item{M}{
  A matrix of numeric type.
}
}
\value{
Returns a vector as the sum of each column in the input matrix.
}
\keyword{internal}
