\name{colsum_Mat_new}
\alias{colsum_Mat_new}
\title{
C++ function for column-wise numeric calculation in a matrix for a weighted static binary undirected network.
}
\description{
The C++ function colsum_Mat_new() calculates the sum of each column in a matrix in a weighted static binary undirected network.
}
\usage{
colsum_Mat_new(M)
}
\arguments{
  \item{M}{
  A matrix of numeric type.
}
}
\value{
Returns a vector as the sum of each column in the input matrix.
}
\keyword{internal}