\name{rowsum_Mat}
\alias{rowsum_Mat}
\title{
C++ function for row-wise numeric calculation in a matrix for a static binary undirected network.
}
\description{
The C++ function rowsum_Mat() calculates the sum of each row in a matrix in a static binary undirected network.
}
\usage{
rowsum_Mat(M)
}
\arguments{
  \item{M}{
  A matrix of numeric type.
}
}
\value{
Returns a vector as the sum of each row in the input matrix.
}
\keyword{internal}