\name{rowsum_Mat_new}
\alias{rowsum_Mat_new}
\title{
C++ function for row-wise numeric calculation in a matrix for a weighted static binary undirected network.
}
\description{
The C++ function rowsum_Mat_new() calculates the sum of each row in a matrix in a weighted static binary undirected network.
}
\usage{
rowsum_Mat_new(M)
}
\arguments{
  \item{M}{
  A matrix of numeric type.
  }
}
\value{
Returns a vector as the sum of each row in the input matrix.
}
\keyword{internal}